updates and reorganization of context apis
