+++ test/react_context_test.dart
@@ -115,9 +115,9 @@ int calculateChangedBits(currentValue, nextValue) {
   return result;
 }
 
-var TestCalculateChangedBitsContext = createContext(1, calculateChangedBits);
+var TestCalculateChangedBitsContext = react.createContext(1, calculateChangedBits);
 
-var TestContext = createContext();
+var TestContext = react.createContext();
 
 ReactDartComponentFactoryProxy2 ContextProviderWrapper = react.registerComponent(() => new _ContextProviderWrapper());
